backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.4.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
anthropic==0.30.0
supabase==2.5.3
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
httpx==0.27.0
aiofiles==23.2.1
nanoid==2.0.0
